requirements-ldap.txt
python-ldap==2.3.13
django-auth-ldap==1.0.12
